+++ inst/htmlwidgets/mapdeck.js
@@ -15,6 +15,15 @@ HTMLWidgets.widget({
 
       	md_setup_window( el.id );
 
+				if( x.show_view_state ) {
+      	  md_setup_view_state( el.id );
+      	  window[el.id + 'mapViewState'] = document.createElement("div");
+      	  window[el.id + 'mapViewState'].setAttribute('id', el.id + 'mapViewState');
+      	  window[el.id + 'mapViewState'].setAttribute('class', 'mapViewState');
+      	  var mapbox_ctrl = document.getElementById( "mapViewStateContainer"+el.id);
+    			mapbox_ctrl.appendChild( window[el.id + 'mapViewState'] );
+				}
+
         /*
         // controller with events
         const myController = new deck.Controller({
@@ -64,7 +73,7 @@ HTMLWidgets.widget({
 			      //onLayerHover: setTooltip
 			      onViewStateChange: ({viewState, interactionState}) => {
 
-			      	if (!HTMLWidgets.shinyMode) { return; }
+			      	if (!HTMLWidgets.shinyMode && !x.show_view_state ) { return; }
 							// as per:
 							// https://github.com/uber/deck.gl/issues/3344
 							// https://github.com/SymbolixAU/mapdeck/issues/211
@@ -78,14 +87,6 @@ HTMLWidgets.widget({
   						const e = se[0] < -180 ? -180 : ( se[0] > 180 ? 180 : se[0] );
   						const s = se[1] < -90 ? -90 : ( se[1] > 90 ? 90 : se[1] );
 
-							/*
-  						console.log(nw);
-  						console.log(se);
-
-  						console.log(w);
-  						console.log(e);
-  						*/
-
   						viewState.viewBounds = {
   							north: n, //nw[1],
   							east:  e, //se[0],
@@ -94,6 +95,13 @@ HTMLWidgets.widget({
   						};
   						viewState.interactionState = interactionState;
 
+  						if( x.show_view_state ) {
+  							var vs = JSON.stringify( viewState );
+  							console.log( vs );
+  							window[el.id + 'mapViewState'].innerHTML = vs;
+  						}
+
+  						if (!HTMLWidgets.shinyMode ) { return; }
 
 						  Shiny.onInputChange(el.id + '_view_change', viewState);
 			      },